feat: Add config file and providers for Neo Link website
This commit adds a new config file (config.ts) to the Neo Link website project, which includes settings for the mainnet and sepolia chains. It also adds a new providers file (providers.tsx) that sets up the QueryClient and WagmiProvider for the application. The providers file also includes a Navbar component that displays the connected account's address and a disconnect button.

The following dependencies were added to the package.json file:
- @tanstack/react-query@^5.56.2
- viem@^2.21.15
- wagmi@^2.12.16

The utils.ts file was also modified to include a new function called shortenAddress, which shortens a given address by truncating the middle part and replacing it with ellipsis.

These changes enhance the functionality and configuration of the Neo Link website.