feat: Add additional parameters and functions to SendPage component
This commit adds additional parameters and functions to the SendPage component in order to enhance its functionality. The changes include importing new functions from the "viem" and "@wagmi/core" libraries, as well as adding new constants and types. The commit also includes the addition of new functions "getChain" and "getLinkForNativeToken" to handle specific chain IDs and generate a link for native tokens. Overall, these changes improve the capabilities of the SendPage component.