feat: Add NeoLinkRaffle contract and related functions
This commit adds the NeoLinkRaffle contract and its related functions to the repository. The NeoLinkRaffle contract allows users to participate in raffles by making deposits using ETH or ERC20 tokens. The batchMakeDeposit and batchMakeDepositRaffle functions handle the deposit process for different contract types. This addition expands the functionality of the project and enhances the user experience.